Add timezoneOffset and datetime types for sleeps
Whoop truncates micro/milliseconds to a number of digits that
isn't 0 modulo 3. Pad it with zeros

whoop seems to fill the timezone substring with UTC +0000 offset
and then sends the actual timezone offset in a separate response field

Take that and patch the timezone substring after padding
to micro/milliseconds.

add tests for that